Create Rescue Dao Coin
THE RESCUE DAO COIN

A decentralized coin created to fund local solidarity projects all over the world.
A tool to transform social capital into financial capital and mobilize a community of support.

The crowdfunding that RESCUE offers involves raising funds by bringing together numerous contributions through the sale of RESCUE and the creation of a solidarity pool.
RESCUE through its website puts the project leader in touch with people who wish to give or invest money.

What is new in the crowdfunding activity is the amplitude reached thanks to the cryptosphere and blockchain technology.
Projects can be proposed here, via the RESCUE website. RESCUE project is for any type of solidarity project that arouses the interest of the community.
An introduction to the world of cryptocurrency is added to the offer to allow the project leader financial autonomy and long-term development.
Finally, through its principle of participation and collective responsibility, the holders of RESCUE can collectively be actors in their own development.
Supporting a RESCUE campaign offers a one-off opportunity for citizen-actors to get involved in this collective development. RESCUE is aware of the fact that we must help. We believe it is our moral obligation and our duty to share with those in need.